Reframe interactions w/ resources as Actions. - Migrated Do"CRUD" operations to instance actions (see instanceaction.go) - Some minor prep for collection actions (will commit separately)
Now resource.Register() accepts an InstanceAction map for a particular resource type. That is put into the resource type's metadata
for future access. When (gomer)gin.BuildRoutes creates a resource hierarchy, allowable routes are determined by what actions are or
are not there. This unblocks "singleton" and other similar use cases.

Data Store.Query() now builds the output items internally rather than via a pre-defined slice of one type. This unblocks the way for
results w/ mixed types (e.g. as you would see when using a single-table design pattern w/ dynamodb). This is accomplished by introduction
of an ItemResolver that accepts an input from the underlying store and turns it into a type suitable for return. The implementations of
these can be provided by users of Gomer, though a default one based on the (currently single-valued) Storable.PersistableTypeName via
a Queryable type handles this for single-typed arrays of a particular Persistable.